Filtering conversational segments of chatbot output.
1) Added the function filter_out_conversational_words, which filters out common english conversational words that ~confuse SD models.

2) Added call to filter_out_conversational_words in the get_SD_pictures function, so that the response from the chatbot is filtered when passed to AUTOMATIC1111's SD api, but still presented unfiltered to the user of oobabooga.
I am submitting this to the last published version of the script on a repo that seems outdated compared to oobabooga's. However, in oobabooga's repo this is mentioned as the experimental branch so, so to not mess things up, the chages here (originally made to the updated script in the obabooga repo) are incorporated in the old script present in the 'Brawlence/SD_api_pics' repo.