A solution temporary but somewhat safisfying solution to too many token taken up by resulting images.
This push fixes the problem of excessive tokens taken up by putting results of SD-generation results taking up too many token.
The solution is very simple: Provide a link (it is a link to cached file in browser, not to one saved to harddrive) to image.

I had to update the old 'not-updated-code' (in comparison to the one in oobabooga), but this enhancement works the same nonetheless.

I recommend updating the code in this repo to at least match the one in oobabooga. Include my enhancement or don't but please update the code in the repo marked as the experimental one; or encourage updates directly in the main oobabooga repo.